return red border around invalid form fields
addresses  #111

The 0-width border in this theme prevented the red error border from showing. This commit returns a 1px solid border with the same CSS selector as PVE's built-in theme and corrects for the corresponding layout disruption by adding a -1px margin. The built-in theme currently sets the border color to #cf4c35, which is not overridden.